Make global_data_all_processes static in run.c.
It now is a local variable containing only parameters needed for the top-level run functions.
